+++ examples/resources/boundary_scope/role.tf
@@ -5,8 +5,8 @@ resource "boundary_scope" "org" {
 }
 
 resource "boundary_role" "org_admin" {
-  scope_id       = boundary_scope.global.id
-  grant_scope_id = boundary_scope.org.id
-  grant_strings  = ["ids=*;type=*;actions=*"]
-  principal_ids  = ["u_auth"]
+  scope_id        = boundary_scope.global.id
+  grant_scope_ids = [boundary_scope.org.id]
+  grant_strings   = ["ids=*;type=*;actions=*"]
+  principal_ids   = ["u_auth"]
 }