+++ internal/provider/data_source_scope_test.go
@@ -42,7 +42,7 @@ resource "boundary_scope" "project" {
 
 resource "boundary_role" "org_admin" {
 	scope_id = "global"
-	grant_scope_id = boundary_scope.org.id
+	grant_scope_ids = [boundary_scope.org.id]
 	grant_strings = ["ids=*;type=*;actions=*"]
 	principal_ids = ["u_auth"]
 }